liblol: add a dummy ld-linux-loongarch-lp64d.so.1
to prevent new world ld.so from loading when an old world executable
loads a new world lib which then requiring the new world ld.so, since
all symbols has been provided by our ld.so.1 in liblol.

This fixes tencent meeting, which loads the new world libstdc++ in the
system, which requires the new world ld.so.